requirements-dev.txt
envparse==0.2.0
Django==1.9.2
django-model-utils==3.0.0